requirements.txt
# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# SQLAlchemy ORM and Database
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0

# Computer Vision and AI
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0

# HTTP and Utilities
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: Development and Testing
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
